Refactor onboarding code and improve exceptions handling
This commit includes the creation of new service layer methods for improving the handling of onboarding processes as well as management of related exceptions. It also renames and updates the representation of a driver's information in DTOs for better clarity. New test cases have been added to ensure the quality of these updates. Gson dependency usage was also corrected in build tools.